Feat run aggregate

+++ domain/repository/queue.go
@@ -6,6 +6,6 @@ import (
 
 type QueueRepository interface {
 	Enqueue(ctx context.Context, key string, members ...any) error
-	Dequeue(ctx context.Context, key string, count int) ([]string, error)
+	Dequeue(ctx context.Context, key string, count int64) ([]string, error)
 	Len(ctx context.Context, key string) (int64, error)
 }
